+++ src/onegov.town/onegov/town/views/userprofile.py
@@ -1,6 +1,8 @@
 """ The settings view, defining things like the logo or color of the town. """
 
-from onegov.core.security import Private
+import morepath
+
+from onegov.core.security import Private, Public
 from onegov.town import _
 from onegov.town.app import TownApp
 from onegov.town.elements import Link
@@ -8,6 +10,7 @@ from onegov.town.forms import UserProfileForm
 from onegov.town.layout import DefaultLayout
 from onegov.town.models import Town
 from onegov.user import UserCollection
+from webob.exc import HTTPForbidden
 
 
 @TownApp.form(
@@ -40,3 +43,51 @@ def handle_user_profile(self, request, form):
         'initials': user.initials,
         'role': user.role
     }
+
+
+# the view name must remain english, so that automated tools can detect it
+@TownApp.html(model=Town, name='unsubscribe', template='userprofile.pt',
+              permission=Public)
+def handle_unsubscribe(self, request):
+    """ Unsubscribes a user from all *regular* e-mails.
+
+    To be able to use this method, an url has to be created like this::
+
+        '{}?token={}'.format((
+            request.link(town, name='unsubscribe'),
+            request.new_url_safe_token(
+                {'user': 'user@example.org'}, 'unsubscribe'
+            )
+        ))
+
+    """
+
+    # tokens are valid for 30 days
+    max_age = 60 * 60 * 24 * 30
+    salt = 'unsubscribe'
+    newsletters = {
+        'daily_ticket_statistics'
+    }
+
+    data = request.load_url_safe_token(
+        request.params.get('token'), max_age=max_age, salt=salt
+    )
+
+    if not data:
+        return HTTPForbidden()
+
+    user = UserCollection(request.app.session()).by_username(data['user'])
+
+    if not user:
+        return HTTPForbidden()
+
+    if not user.data:
+        user.data = {}
+
+    for newsletter in newsletters:
+        user.data[newsletter] = False
+
+    request.success(
+        _("You have been successfully unsubscribed from all regular emails.")
+    )
+    return morepath.redirect(request.link(self))